add land_commodity service

+++ internal/model/domain/land _entity.go
@@ -11,7 +11,7 @@ type Land struct {
 	ID          uuid.UUID      `gorm:"primaryKey;type:varchar(255)"`
 	UserID      uuid.UUID      `gorm:"not null;type:varchar(255)"`
 	User        User           `gorm:"foreignKey:UserID" json:"-"`
-	LandArea    int64          `gorm:"not null;type:bigint"`
+	LandArea    float64        `gorm:"not null;type:bigint"`
 	Certificate string         `gorm:"not null;type:varchar(255)"`
 	CreatedAt   time.Time      `gorm:"autoCreateTime"`
 	UpdatedAt   time.Time      `gorm:"autoUpdateTime"`
